requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-phone-field==1.8.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
PyJWT==1.7.1
